pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
authors = [
]
dependencies = [
    "django-extensions>=3.2.3",
    "djangorestframework>=3.14.0",
    "Django>=4.2.7",
    "pymysql>=1.1.0",
    "dj_database_url>=2.1.0",
    "geopy>=2.4.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=23.10.1",
]

[tool.hatch.metadata]
allow-direct-references = true
